# assignments/assign03-serverless/docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Serverless Order Processing API
  description: |
    REST API for serverless e-commerce order processing system.
    This API provides endpoints for creating and retrieving orders in a fully serverless architecture.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://api.example.com/prod
    description: Production server
  - url: https://api.example.com/dev
    description: Development server

tags:
  - name: Orders
    description: Order management operations

security:
  - ApiKeyAuth: []

paths:
  /orders:
    post:
      tags:
        - Orders
      summary: Create a new order
      description: Creates a new order and submits it for processing
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              basic:
                summary: Basic order
                value:
                  customerId: "CUST-12345"
                  customerEmail: "customer@example.com"
                  items:
                    - productId: "PROD-001"
                      quantity: 2
                      price: 29.99
                    - productId: "PROD-002"
                      quantity: 1
                      price: 49.99
                  shippingAddress:
                    street: "123 Main St"
                    city: "New York"
                    state: "NY"
                    zipCode: "10001"
                    country: "USA"
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                message: "Order created successfully"
                orderId: "550e8400-e29b-41d4-a716-446655440000"
                status: "PENDING"
                totalAmount: 109.97
                createdAt: "2024-01-15T10:30:00Z"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  value:
                    error: "ValidationError"
                    message: "Missing required field: customerId"
                invalidJSON:
                  value:
                    error: "InvalidJSON"
                    message: "Request body must be valid JSON"
        '401':
          description: Unauthorized - Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Orders
      summary: List all orders
      description: Retrieves a list of orders, optionally filtered by status
      operationId: listOrders
      parameters:
        - name: status
          in: query
          description: Filter orders by status
          required: false
          schema:
            type: string
            enum: [PENDING, PROCESSING, COMPLETED, FAILED]
          example: PENDING
        - name: limit
          in: query
          description: Maximum number of orders to return (max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          example: 20
      responses:
        '200':
          description: List of orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
              example:
                orders:
                  - orderId: "550e8400-e29b-41d4-a716-446655440000"
                    customerId: "CUST-12345"
                    status: "PENDING"
                    totalAmount: 109.97
                    createdAt: "2024-01-15T10:30:00Z"
                  - orderId: "550e8400-e29b-41d4-a716-446655440001"
                    customerId: "CUST-67890"
                    status: "COMPLETED"
                    totalAmount: 75.50
                    createdAt: "2024-01-15T09:15:00Z"
                count: 2
                filter:
                  status: null
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Retrieves detailed information about a specific order
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetails'
              example:
                orderId: "550e8400-e29b-41d4-a716-446655440000"
                createdAt: "2024-01-15T10:30:00Z"
                customerId: "CUST-12345"
                customerEmail: "customer@example.com"
                items:
                  - productId: "PROD-001"
                    quantity: 2
                    price: 29.99
                  - productId: "PROD-002"
                    quantity: 1
                    price: 49.99
                totalAmount: 109.97
                status: "PENDING"
                updatedAt: "2024-01-15T10:30:00Z"
                shippingAddress:
                  street: "123 Main St"
                  city: "New York"
                  state: "NY"
                  zipCode: "10001"
                  country: "USA"
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "NotFound"
                message: "Order 550e8400-e29b-41d4-a716-446655440000 not found"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

  schemas:
    CreateOrderRequest:
      type: object
      required:
        - customerId
        - items
      properties:
        customerId:
          type: string
          description: Unique customer identifier
          example: "CUST-12345"
        customerEmail:
          type: string
          format: email
          description: Customer email address for notifications
          example: "customer@example.com"
        items:
          type: array
          description: List of items in the order
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/Address'

    OrderItem:
      type: object
      required:
        - productId
        - quantity
        - price
      properties:
        productId:
          type: string
          description: Unique product identifier
          example: "PROD-001"
        quantity:
          type: integer
          minimum: 1
          description: Quantity of the product
          example: 2
        price:
          type: number
          format: float
          minimum: 0
          description: Price per unit
          example: 29.99

    Address:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        state:
          type: string
          example: "NY"
        zipCode:
          type: string
          example: "10001"
        country:
          type: string
          example: "USA"

    OrderResponse:
      type: object
      properties:
        message:
          type: string
          example: "Order created successfully"
        orderId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
          example: "PENDING"
        totalAmount:
          type: number
          format: float
          example: 109.97
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"

    OrderDetails:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        customerId:
          type: string
        customerEmail:
          type: string
          format: email
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: float
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        updatedAt:
          type: string
          format: date-time
        shippingAddress:
          $ref: '#/components/schemas/Address'

    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/OrderSummary'
        count:
          type: integer
          description: Number of orders returned
        filter:
          type: object
          nullable: true
          properties:
            status:
              type: string
              nullable: true

    OrderSummary:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        customerId:
          type: string
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        totalAmount:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Human-readable error message
          example: "Missing required field: customerId"